requirements.txt
flask==2.0.1
pandas==1.3.3
openpyxl==3.0.9
python-dotenv==0.19.0
requests==2.26.0 